refactor: move all cache references from graphql pkg to cache pkg
